requirements-docs.txt
# Documentation agents dependencies
agent-framework-azure-ai==1.0.0b260107
agent-framework-core==1.0.0b260107
azure-identity==1.24.0
python-dotenv==1.0.1
